rename files and add readme for llama auto_parallel
