+++ scripts/distribute/ci_case_auto.sh
@@ -1252,7 +1252,7 @@ function llama_dygraph_auto_bs8_fp32_DP2() {
     rm -rf $case_out_dir
     rm -rf $case_log_dir
 
-    python -u -m paddle.distributed.launch --gpus "0,1" --log_dir $case_log_dir run_pretrain_3D_auto.py \
+    python -u -m paddle.distributed.launch --gpus "0,1" --log_dir $case_log_dir run_pretrain_auto.py \
         --model_type "llama" \
         --model_name_or_path "facebook/llama-7b" \
         --tokenizer_name_or_path "facebook/llama-7b" \
@@ -1319,7 +1319,7 @@ function llama_dygraph_auto_bs8_fp32_DP2-MP2() {
     rm -rf $case_out_dir
     rm -rf $case_log_dir
 
-    python -u -m paddle.distributed.launch --gpus "0,1,2,3" --log_dir $case_log_dir run_pretrain_3D_auto.py \
+    python -u -m paddle.distributed.launch --gpus "0,1,2,3" --log_dir $case_log_dir run_pretrain_auto.py \
         --model_type "llama" \
         --model_name_or_path "facebook/llama-7b" \
         --tokenizer_name_or_path "facebook/llama-7b" \
@@ -1386,7 +1386,7 @@ function llama_dygraph_auto_bs8_fp32_DP2-MP2-PP2() {
     rm -rf $case_out_dir
     rm -rf $case_log_dir
 
-    python -u -m paddle.distributed.launch --gpus "0,1,2,3,4,5,6,7" --log_dir $case_log_dir run_pretrain_3D_auto.py \
+    python -u -m paddle.distributed.launch --gpus "0,1,2,3,4,5,6,7" --log_dir $case_log_dir run_pretrain_auto.py \
         --model_type "llama" \
         --model_name_or_path "facebook/llama-7b" \
         --tokenizer_name_or_path "facebook/llama-7b" \
